SearchService/requirements.txt
fastapi==0.115.12
grpcio==1.71.0
grpcio-tools==1.71.0
pika==1.3.2
pinecone==6.0.2
pydantic==2.11.3
python-dotenv==1.1.0
sentence-transformers==4.0.2
uvicorn==0.34.0
huggingface-hub==0.30.2
pytest==8.3.5
httpx==0.28.1
hf-xet==1.0.2
SentencePiece==0.2.0
tiktoken==0.9.0
confluent-kafka==2.10.0
geoip2==5.1.0